+++ hexital/analysis/utils.py
@@ -43,6 +43,22 @@ def shadow_lower_avg(candles: List[Candle], length: int, index: Optional[int] = 
     return sum(candles[i].shadow_lower for i in range(start_index, index)) / length
 
 
+def realbody_gapdown(candle: Candle, candle_two: Candle) -> bool:
+    return min(candle.open, candle.close) > max(candle_two.open, candle_two.close)
+
+
+def realbody_gapup(candle: Candle, candle_two: Candle) -> bool:
+    return max(candle.open, candle.close) < min(candle_two.open, candle_two.close)
+
+
+def candle_gapup(candle: Candle, candle_two: Candle) -> bool:
+    return candle.low > candle_two.high
+
+
+def candle_gapdown(candle: Candle, candle_two: Candle) -> bool:
+    return candle.high < candle_two.low
+
+
 # Below are TA-lib globally used utils
 # https://github.com/TA-Lib/ta-lib/blob/main/src/ta_common/ta_global.c